+++ ldapsync/tests/test_clone.py
@@ -1,15 +1,13 @@
-from datetime import date, datetime
+from datetime import datetime, date
 
 from django.db.models import Q
 from django.test import TestCase
 from django.utils import timezone
 
-from ldapsync.clone import CloneError, clone
-from ldapsync.ldapoperations import AddOperation, DeleteOperation, ModifyDNOperation, ModifyOperation
-from ldapsync.models import LDAPGroup
-from ldapsync.sync import sync
-from members.models import Person, QGroup, ExternalCard, ExternalCardLoan, Instrument, GSuiteAccount, Key, \
-    GroupMembership
+from ldapsync.clone import clone, CloneError
+from ldapsync.ldapoperations import ModifyOperation
+from members.models import Person, QGroup, GroupMembership, Instrument, ExternalCard, ExternalCardLoan, GSuiteAccount, \
+    Key
 
 
 class CloneTestCase(TestCase):
@@ -368,86 +366,3 @@ class CloneTestCase(TestCase):
                   ModifyOperation('cn=huidige leden,ou=groups,dc=esmgquadrivium,dc=nl', 'linkID', [group1_id]),
                   ModifyOperation('cn=agroup,ou=groups,dc=esmgquadrivium,dc=nl', 'linkID', [group2_id])]
         self.assertCountEqual(expect, actual)
-
-
-class SyncTestCase(TestCase):
-    """Test one-way sync of two datasets."""
-
-    def test_none(self):
-        """No operations need to be performed."""
-        source = {'uid=peep': {'uid': ['peep'], 'name': ['Henk'], 'id': [4]}}
-        target = {'uid=peep': {'uid': ['peep'], 'name': ['Henk'], 'id': [4]}}
-        operations = sync(source, target, on='id')
-        self.assertEqual([], operations)
-
-    def test_add(self):
-        source = {'uid=test': {'uid': ['test'], 'id': ['4']}}
-        target = {}
-        operations = sync(source, target, on='id')
-        expect = [AddOperation('uid=test', {'uid': ['test'], 'id': ['4']})]
-        self.assertEqual(expect, operations)
-
-    def test_delete(self):
-        source = {}
-        target = {'uid=test': {'uid': ['test'], 'id': ['4']}}
-        operations = sync(source, target, on='id')
-        self.assertEqual([DeleteOperation('uid=test')], operations)
-
-    def test_modify(self):
-        change_to = {'uid=test': {'uid': ['test'], 'name': ['Piet'], 'id': ['4']}}
-        to_change = {'uid=test': {'uid': ['test'], 'name': ['Henk'], 'id': ['4']}}
-        operations = sync(change_to, to_change, on='id')
-        self.assertEqual([ModifyOperation('uid=test', 'name', ['Piet'])], operations)
-
-    def test_modify_list(self):
-        change_to = {'uid=test': {'uid': ['test'], 'name': ['Piet', 'Henk'], 'id': [4]}}
-        to_change = {'uid=test': {'uid': ['test'], 'name': ['Piet'], 'id': [4]}}
-        operations = sync(change_to, to_change, on='id')
-        self.assertEqual([ModifyOperation('uid=test', 'name', ['Piet', 'Henk'])], operations)
-
-    def test_modify_list2(self):
-        source = {'uid=test': {'uid': ['test'], 'name': ['Henk', 'Piet'], 'id': [4]}}
-        target = {'uid=test': {'uid': ['test'], 'name': ['Piet', 'Henk'], 'id': [4]}}
-        operations = sync(source, target, on='id')
-        self.assertEqual([], operations)
-
-    def test_modify_dn(self):
-        change_to = {'uid=test': {'id': ['4']}}
-        to_change = {'uid=test2': {'id': ['4']}}
-        operations = sync(change_to, to_change, on='id')
-        self.assertEqual([ModifyDNOperation('uid=test2', 'uid=test')], operations)
-
-    def test_modify_dn_attribute(self):
-        change_to = {'uid=test': {'uid': ['test'], 'id': ['4']}}
-        to_change = {'uid=test2': {'uid': ['test2'], 'id': ['4']}}
-        operations = sync(change_to, to_change, on='id')
-        # ModifyDNOperation should be before ModifyOperation
-        self.assertEqual([ModifyDNOperation('uid=test2', 'uid=test'),
-                          ModifyOperation('uid=test', 'uid', ['test'])], operations)
-
-    def test_entry_without_matching_attribute(self):
-        change_to = {'uid=test': {'uid': ['test'], 'id': [1]}}
-        to_change1 = {'uid=test': {'uid': ['test']}}
-        to_change2 = {'uid=test': {'uid': ['test'], 'id': []}}
-        operations1 = sync(change_to, to_change1, on='id')
-        operations2 = sync(change_to, to_change2, on='id')
-        expect = [DeleteOperation('uid=test'), AddOperation('uid=test', {'uid': ['test'], 'id': [1]})]
-        self.assertEqual(expect, operations1)
-        self.assertEqual(expect, operations2)
-
-
-class ModelsTestCase(TestCase):
-    def test_group_membership_current(self):
-        """Test group membership is included in member list."""
-        p = Person.objects.create(username='test')
-        g = QGroup.objects.create(name='group')
-        p.groups.add(g)
-        ldap_attrs = LDAPGroup.objects.first().get_attributes()
-        self.assertEqual(['uid=test,ou=people,dc=esmgquadrivium,dc=nl'], ldap_attrs['member'])
-
-    def test_no_group_membership(self):
-        """Test not a group member."""
-        Person.objects.create(username='test')
-        QGroup.objects.create(name='group')
-        ldap_attrs = LDAPGroup.objects.first().get_attributes()
-        self.assertEqual([], ldap_attrs['member'])
