+++ ldapsync/tests/test_models.py
@@ -1,0 +1,21 @@
+from django.test import TestCase
+
+from ldapsync.models import LDAPGroup
+from members.models import Person, QGroup
+
+
+class ModelsTestCase(TestCase):
+    def test_group_membership_current(self):
+        """Test group membership is included in member list."""
+        p = Person.objects.create(username='test')
+        g = QGroup.objects.create(name='group')
+        p.groups.add(g)
+        ldap_attrs = LDAPGroup.objects.first().get_attributes()
+        self.assertEqual(['uid=test,ou=people,dc=esmgquadrivium,dc=nl'], ldap_attrs['member'])
+
+    def test_no_group_membership(self):
+        """Test not a group member."""
+        Person.objects.create(username='test')
+        QGroup.objects.create(name='group')
+        ldap_attrs = LDAPGroup.objects.first().get_attributes()
+        self.assertEqual([], ldap_attrs['member'])
